requirements.txt
canvasapi>=2.1.0
gkeepapi>=0.13.4
pyyaml>=5.3.1
